+++ cmd/gravinet/main.go
@@ -46,7 +46,7 @@ import (
 
 // Build metadata, overridable via -ldflags.
 var (
-	version = "453"
+	version = "455"
 	commit  = "none"
 )